[refactor] domFor: split `delayedRemoval` into a separate file
The bundle size can be reduced by removing the intermediate object used to export domFor and delayedRemoval.
Additionally, the file paths used in domFor-related require() calls have been standardized.
