+++ index.js
@@ -3,7 +3,6 @@
 var hyperscript = require("./hyperscript")
 var request = require("./request")
 var mountRedraw = require("./mount-redraw")
-var domFor = require("./render/domFor")
 
 var m = function m() { return hyperscript.apply(this, arguments) }
 m.m = hyperscript
@@ -21,6 +20,6 @@ m.parsePathname = require("./pathname/parse")
 m.buildPathname = require("./pathname/build")
 m.vnode = require("./render/vnode")
 m.censor = require("./util/censor")
-m.domFor = domFor.domFor
+m.domFor = require("./render/domFor")
 
 module.exports = m
